Move pages from components to separate pages folder
