+++ src/pages/AboutPage.jsx
@@ -8,7 +8,6 @@ const AboutPage = () => {
         <p className="text-lg text-gray-700 mb-6">
           Merry Berry is dedicated to serving the finest smoothies and bowls made from 100% organic ingredients. We believe in fresh flavors, sustainable sourcing, and a vibrant community spirit. Our menu is crafted with care to provide a healthy, delicious experience in every bite.
         </p>
-        {/* ...additional content or components... */}
       </div>
     </div>
   );
