+++ src/pages/ContactPage.jsx
@@ -1,0 +1,58 @@
+import React from 'react';
+import { toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
+import { Formik, Form, Field, ErrorMessage } from 'formik';
+import * as Yup from 'yup';
+import Layout from '../components/Layout';
+
+const ContactPage = () => {
+  const validationSchema = Yup.object({
+    name: Yup.string().required('Name is required'),
+    email: Yup.string().email('Invalid email address').required('Email is required'),
+    message: Yup.string().required('Message is required'),
+  });
+
+  const handleSubmit = (values, { setSubmitting, resetForm }) => {
+    toast.success('Message sent successfully!');
+    resetForm();
+    setSubmitting(false);
+  };
+
+  return (
+    <Layout>
+      <div className="max-w-md mx-auto mt-10 p-5 border rounded shadow-lg">
+        <h1 className="text-2xl font-bold mb-5">Contact Us</h1>
+        <Formik
+          initialValues={{ name: '', email: '', message: '' }}
+          validationSchema={validationSchema}
+          onSubmit={handleSubmit}
+        >
+          {({ isSubmitting }) => (
+            <Form className="space-y-4">
+              <div>
+                <label className="block text-sm font-medium text-gray-700">Name:</label>
+                <Field type="text" name="name" className="mt-1 block w-full p-2 border border-gray-300 rounded-md shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm" />
+                <ErrorMessage name="name" component="div" className="text-red-500 text-sm mt-1" />
+              </div>
+              <div>
+                <label className="block text-sm font-medium text-gray-700">Email:</label>
+                <Field type="email" name="email" className="mt-1 block w-full p-2 border border-gray-300 rounded-md shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm" />
+                <ErrorMessage name="email" component="div" className="text-red-500 text-sm mt-1" />
+              </div>
+              <div>
+                <label className="block text-sm font-medium text-gray-700">Message:</label>
+                <Field as="textarea" name="message" className="mt-1 block w-full p-2 border border-gray-300 rounded-md shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm" />
+                <ErrorMessage name="message" component="div" className="text-red-500 text-sm mt-1" />
+              </div>
+              <button type="submit" disabled={isSubmitting} className="w-full bg-indigo-600 text-white p-2 rounded-md shadow-sm hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500">
+                Send
+              </button>
+            </Form>
+          )}
+        </Formik>
+      </div>
+    </Layout>
+  );
+};
+
+export default ContactPage;
